requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
openai>=1.0.0
anthropic>=0.25.0
pinecone>=3.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
notion-client>=2.2.0
python-dotenv>=1.0.0
